pkg/endpoint: move error types to errors.go
Signed-off by: Ian Vernon <ian@cilium.io>
