+++ pkg/endpoint/errors.go
@@ -1,0 +1,36 @@
+// Copyright 2018 Authors of Cilium
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package endpoint
+
+type UpdateValidationError struct {
+	msg string
+}
+
+func (e UpdateValidationError) Error() string { return e.msg }
+
+type UpdateCompilationError struct {
+	msg string
+}
+
+func (e UpdateCompilationError) Error() string { return e.msg }
+
+// UpdateStateChangeError is an error that indicates that updating the state
+// of an endpoint was unsuccessful.
+// Implements error interface.
+type UpdateStateChangeError struct {
+	msg string
+}
+
+func (e UpdateStateChangeError) Error() string { return e.msg }
